refactor FormCandidateField and FIELD_TYPES const
