+++ src/consts/fieldTypes.js
@@ -1,0 +1,11 @@
+const FIELD_TYPES = {
+  text: 'text',
+  textarea: 'textarea',
+  radio: 'radio',
+  radioGroup: 'radioGroup',
+  phone: 'phone',
+  email: 'email',
+  date: 'date',
+};
+
+export default FIELD_TYPES;
