+++ src/consts/index.js
@@ -1,0 +1,5 @@
+import FIELD_TYPES from "./fieldTypes";
+
+export {
+  FIELD_TYPES,
+};
