+++ src/containers/FormCandidate/FormCandidateField/FormCandidateField.jsx
@@ -1,7 +1,48 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 
+import { FIELD_TYPES } from '../../../consts';
 import CustomTextField from '../../../components/@ui/CustomTextField';
 import useApp from '../hooks/useApp';
+
+const {
+  text,
+  textarea,
+  radio,
+  radioGroup,
+  phone,
+  email,
+  date,
+} = FIELD_TYPES;
+
+const DisplayField = ({
+  field,
+  onChange,
+}) => {
+  const {
+    title,
+    code,
+    type,
+    col,
+    required,
+  } = field;
+
+  switch (type) {
+    case text:
+      return <CustomTextField
+        label={field.title}
+        value={field.value}
+        onChange={onChange}
+      />
+
+    default:
+      return <CustomTextField
+        label={field.title}
+        value={field.value}
+        onChange={onChange}
+      />
+  }
+}
 
 const FormCandidateField = ({
   field,
@@ -12,14 +53,35 @@ const FormCandidateField = ({
     handleFieldChange,
   } = useApp();
 
+  const handleDisplayFieldChange = (value) => {
+    handleFieldChange(sectionCode, fieldCode, value)
+  }
+
   return (
-    <CustomTextField
-      key={fieldCode}
-      label={field.title}
-      value={field.value}
-      onChange={(value) => handleFieldChange(sectionCode, fieldCode, value)}
+    <DisplayField
+      field={field}
+      onChange={handleDisplayFieldChange}
     />
   );
+};
+
+FormCandidateField.propTypes = {
+  field: PropTypes.shape({
+    title: PropTypes.string.isRequired,
+    code: PropTypes.string,
+    type: PropTypes.oneOf(Object.keys(FIELD_TYPES)).isRequired,
+    col: PropTypes.number,
+    required: PropTypes.bool,
+  }),
+};
+DisplayField.propTypes = {
+  field: PropTypes.shape({
+    title: PropTypes.string.isRequired,
+    code: PropTypes.string,
+    type: PropTypes.oneOf(Object.keys(FIELD_TYPES)).isRequired,
+    col: PropTypes.number,
+    required: PropTypes.bool,
+  }),
 };
 
 export default FormCandidateField;
