+++ src/redux/actions/model/Form/dto.js
@@ -1,12 +1,4 @@
-const FIELD_TYPES = {
-  text: 'text',
-  textarea: 'textarea',
-  radio: 'radio',
-  radioGroup: 'radioGroup',
-  phone: 'phone',
-  email: 'email',
-  date: 'date',
-};
+import { FIELD_TYPES } from "../../../../consts";
 
 const {
   text,
